--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -103,6 +103,7 @@ Collate:
     'utilities-path.R'
     'standard-path.R'
     'table-formula.R'
+    'tasks-env.R'
     'user-defined-pk-parameter.R'
     'utilities-container.R'
     'utilities-data-combined.R'