--- a/R/data-repository.R
+++ b/R/data-repository.R
@@ -93,7 +93,7 @@ DataRepository <- R6::R6Class(
       }
       validateIsString(name)
       validateIsString(value)
-      dataRepositoryTask <- .getNetTask("DataRepositoryTask")
+      dataRepositoryTask <- .getNetTaskFromCache("DataRepositoryTask")
       rClr::clrCall(dataRepositoryTask, "AddMetaData", self$ref, name, value)
       # we need to reset the cache when adding a new meta data
       private$.metaData <- NULL
@@ -107,7 +107,7 @@ DataRepository <- R6::R6Class(
         stop(messages$errorMultipleMetaDataEntries())
       }
       validateIsString(name)
-      dataRepositoryTask <- .getNetTask("DataRepositoryTask")
+      dataRepositoryTask <- .getNetTaskFromCache("DataRepositoryTask")
       rClr::clrCall(dataRepositoryTask, "RemoveMetaData", self$ref, name)
       # we need to reset the cache when adding a new meta data
       private$.metaData <- NULL