--- a/R/get-net-task.R
+++ b/R/get-net-task.R
@@ -1,4 +1,4 @@
-#' @title .getNewNetTask
+#' @title .getNetTask
 #' @description Get an instance of the specified `.NET` Task
 #'
 #' @param taskName The name of the task to retrieve (**without** `Get` prefix).
@@ -6,11 +6,11 @@
 #' @return returns a new instance of of the specified `.NET` task.
 #'
 #' @keywords internal
-.getNewNetTask <- function(taskName) {
+.getNetTask <- function(taskName) {
   rClr::clrCallStatic("OSPSuite.R.Api", paste0("Get", taskName))
 }
 
-#' @title .getNetTask
+#' @title .getNetTaskFromCache
 #' @description Get an instance of the specified `.NET` Task that is retrieved
 #' from cache if already initiated. Otherwise a new task will be initiated and
 #' cached in the `tasksEnv`.
@@ -20,25 +20,9 @@
 #' @return returns an instance of of the specified `.NET` task.
 #'
 #' @keywords internal
-.getNetTask <- function(taskName) {
+.getNetTaskFromCache <- function(taskName) {
   if (is.null(tasksEnv[[taskName]])) {
-    tasksEnv[[taskName]] <- .getNewNetTask(taskName)
+    tasksEnv[[taskName]] <- .getNetTask(taskName)
   }
   return(tasksEnv[[taskName]])
-}
-
-#' @title .getDimensionTask
-#'
-#' @return an instance of the `.NET` Task `DimensionTask`
-#' @keywords internal
-.getDimensionTask <- function() {
-  .getNetTask("DimensionTask")
-}
-
-#' @title .getContainerTask
-#'
-#' @return an instance of the `.NET` Task `ContainerTask`
-#' @keywords internal
-.getContainerTask <- function() {
-  .getNetTask("ContainerTask")
 }