--- a/R/tasks-env.R
+++ b/R/tasks-env.R
@@ -1,0 +1,3 @@
+# Environment that holds .Net tasks initiated by ospsuite,
+# It is not exported and should not be directly manipulated by other packages.
+tasksEnv <- new.env(parent = emptyenv())