--- a/R/utilities-data-importer-configuration.R
+++ b/R/utilities-data-importer-configuration.R
@@ -22,7 +22,7 @@
 #' @export
 createImporterConfigurationForFile <- function(filePath, sheet = NULL) {
   validateIsString(filePath)
-  importerTask <- .getNetTask("DataImporterTask")
+  importerTask <- .getNetTaskFromCache("DataImporterTask")
   if (is.null(sheet)) {
     ref <- rClr::clrCall(importerTask, "CreateConfigurationFor", filePath)
   } else {
@@ -60,7 +60,7 @@ createImporterConfigurationForFile <- function(filePath, sheet = NULL) {
 #' @export
 loadDataImporterConfiguration <- function(configurationFilePath) {
   validateIsString(configurationFilePath)
-  importerTask <- .getNetTask("DataImporterTask")
+  importerTask <- .getNetTaskFromCache("DataImporterTask")
   ref <- rClr::clrCall(importerTask, "GetConfiguration", configurationFilePath)
   return(DataImporterConfiguration$new(ref))
 }