--- a/R/utilities-simulation-results.R
+++ b/R/utilities-simulation-results.R
@@ -109,7 +109,7 @@ getOutputValues <- function(simulationResults,
   names(values) <- paths
 
   # Use low-level methods to get unit and dimension
-  task <- .getContainerTask()
+  task <- .getNetTaskFromCache("ContainerTask")
   metaData <- NULL
   if (addMetaData) {
     metaData <- lapply(paths, function(path) {