--- a/R/utilities-units.R
+++ b/R/utilities-units.R
@@ -5,7 +5,7 @@
 #' @export
 hasDimension <- function(dimension) {
   validateIsString(dimension)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "HasDimension", enc2utf8(dimension))
 }
 
@@ -30,7 +30,7 @@ validateDimension <- function(dimension) {
 hasUnit <- function(unit, dimension) {
   validateIsString(unit)
   validateDimension(dimension)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "HasUnit", enc2utf8(dimension), .encodeUnit(unit))
 }
 
@@ -71,7 +71,7 @@ validateUnit <- function(unit, dimension) {
 #' @export
 getBaseUnit <- function(dimension) {
   validateDimension(dimension)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "BaseUnitFor", enc2utf8(dimension))
 }
 
@@ -180,7 +180,7 @@ toUnit <- function(quantityOrDimension,
     molWeight <- NULL
   }
 
-  dimensionTask <- .getNetTask("DimensionTask")
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   # ensure that we are dealing with an list of values seen as number (and not integer)
   values <- as.numeric(c(values))
 
@@ -252,7 +252,7 @@ toDisplayUnit <- function(quantity, values) {
 #' allAvailableDimensions()
 #' @export
 allAvailableDimensions <- function() {
-  dimensionTask <- .getNetTask("DimensionTask")
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "AllAvailableDimensionNames")
 }
 
@@ -272,7 +272,7 @@ allAvailableDimensions <- function() {
 getDimensionForUnit <- function(unit) {
   validateIsString(unit)
   unit <- .encodeUnit(unit)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   dim <- rClr::clrCall(dimensionTask, "DimensionForUnit", unit)
   ifNotNull(dim, rClr::clrGet(dim, "Name"))
 }
@@ -291,7 +291,7 @@ getDimensionForUnit <- function(unit) {
 #' @export
 getUnitsForDimension <- function(dimension) {
   validateIsString(dimension)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "AllAvailableUnitNamesFor", enc2utf8(dimension))
 }
 
@@ -310,7 +310,7 @@ getUnitsForDimension <- function(dimension) {
 #' @export
 getDimensionByName <- function(name) {
   validateIsString(name)
-  dimensionTask <- .getDimensionTask()
+  dimensionTask <- .getNetTaskFromCache("DimensionTask")
   rClr::clrCall(dimensionTask, "DimensionByName", enc2utf8(name))
 }