## Summary
While profiling the code for performance improvement, I noticed that it was actually the `.getNetTask` function that was called thousand times by the `toUnit` function, even for "simple" plots.
What  I propose here is a way to cache .Net Task in the package environment the first time it is created.

## Symptom 🩺
Plots can take very long time to generate. 

## Problem 🪲

This is what I saw during profiling of a project's plot:

![Pasted image 20230407115430](https://user-images.githubusercontent.com/34234913/230639489-67938736-1a63-4884-b9bb-fc48072c04e1.png)

Here I focus on the **`toUnit()` function that took 6.6 seconds** for a plot.

When tested on its own, this function is actually quick, even with large numeric vectors. So I counted the number of time this function was called to generate **1 plotGrid**... **8934** times !

Let's try to call the function 8934 times for converting 1 vallue from L to mL:
```r
> simPath <- system.file("extdata", "simple.pkml", package = "ospsuite")
> sim <- loadSimulation(simPath)
> par <- getParameter("Organism|Liver|Volume", sim)
> bench::mark(toUnit(par, 1, "ml"), iterations = 8934)
```
```
# A tibble: 1 × 13
  expression                min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result    memory              time               gc                  
  <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>    <list>              <list>             <list>              
1 toUnit(par, 1, "ml")   1.13ms   1.17ms      731.    47.3KB     2.38  8905    29      12.2s <dbl [1]> <Rprofmem [47 × 3]> <bench_tm [8,934]> <tibble [8,934 × 3]>
```

**12 seconds**

But the culprit is actually `.getNetTask("DimensionTask")` which is called by `toUnit()`:
```
> bench::mark(.getNetTask(.getNetTask("DimensionTask")), iterations = 8934)
# A tibble: 1 × 13
  expression                        min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result    memory             time               gc                  
  <bch:expr>                   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>    <list>             <list>             <list>              
1 .getNetTask("DimensionTask")    827µs    847µs     1023.      328B    0.803  8927     7      8.72s <cobjRef> <Rprofmem [1 × 3]> <bench_tm [8,934]> <tibble [8,934 × 3]>
```
**8.72 seconds !**

So, **creating new "DimensionTask" repeatedly is what takes time** in this part of the code.

## Proposition 💡

Create .Net Tasks just once: when we need them for the first time. Then, they are stored in the `ospsuiteEnv` and reused. This is not true for Simulation tasks, a new one is created for each simulation.

## Results 🚀

➡️  **82 % less time for the toUnit function**

```
bench::mark(toUnit(par, 1, "ml"), toUnit_old(par, 1, "ml"), iterations = 8900)
# A tibble: 2 × 13
  expression                    min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result    memory               time               gc                  
  <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>    <list>               <list>             <list>              
1 toUnit(par, 1, "ml")        208µs  226.7µs     3458.     254KB    10.9   8872    28      2.56s <dbl [1]> <Rprofmem [522 × 3]> <bench_tm [8,900]> <tibble [8,900 × 3]>
2 toUnit_old(par, 1, "ml")   1.14ms    1.2ms      741.     205KB     3.51  8858    42     11.95s <dbl [1]> <Rprofmem [474 × 3]> <bench_tm [8,900]> <tibble [8,900 × 3]>
```

![image](https://user-images.githubusercontent.com/34234913/230641872-a49345c4-c48c-4d7a-b484-5ed5106aef68.png)

The plot I used to debug this still takes time: **16s but previously it was 24s**! The biggest part is still to improve: `private$wrapExtensionMethod` from `data-column.R`  takes a total of 8 seconds and I'm sure we can do something about it.

## Quality ✅

```r
══ Results ═══════════════════════════════════════════════════════════════════════
Duration: 191.4 s

[ FAIL 0 | WARN 10 | SKIP 0 | PASS 1232 ]
```

## Potential side effects ❓
I am not familiar with the Task principle on the `.Net` side. Maybe @georgeDaskalakis  & @rwmcintosh have any comment on this ?