@PavelBal 2023-04-07
This is great and I think it should be cafe to cache net tasks. I would like @msevestre to review to be sure.