@Felixmil 2023-04-11
Thanks for the extended review.

1. It does not seem to be the case for simulation tasks (see comment on 3 below)
2. Fixed in last commit
3. > 3. Why rename from get to create? This is semantically wrong as I am expecting a new task to be created when I call a create method. In this case, we are actually not doing this at all and always caching => rename back please

    Originaly, I just changed `.getNetTask()` to add the implemented logic BUT it led to errors for simulation tasks:
    ```
     ── Error ('test-utilities-simulation.R:290:3'): It can run a simulation batch by varying some parameters and molecules ──
    Error in `rClr::clrCall(simulationRunner, "RunConcurrently")`: Type:    OSPSuite.Utility.Exceptions.OSPSuiteException
    Message: You already have Simulation and SimulationBatch objects and should not mix, please invoke Clear to start adding 
    objects from a fresh start
    ```
    So I needed to start simulation objects "from fresh start"  by running the original `.getNetTask`  each time it would be needed     (with `.createNetTask`) just like it was before. Maybe `getNewNetTask` would be better naming ?
4. Fixed in last commit