@msevestre 2023-04-11
I think we need to make sure we return multiple instances and not only one. Creating two instances in R or .net is probably not what we need but let me check it out and I'll let you know