@msevestre 2023-04-14
@Felixmil is looking at the dimensiontask that can be cached. I am talking about the ConcurrentSimulationRunner. Creating time is going to be << running time so in this case, this caching is not required. I am just wondering why this crashes in fact and I think that those tasks are not stateless in 
.NET. I had no time to look at this but I will to understand at least
As long as the PR doesn't cache them all , we should be good. 
In fact, global caching is the source of many evil. I'll suggest to only cache when we have too as opposed to caching all the time . DimensionTask, ContainerTask are good candidates.