@PavelBal 2023-04-14
We are creating the following tasks:

- `DataRepositoryTask` (propose to cache)
- `DataImporterTask` (propose to cache)
- `OutputIntervalFactory`
- `PKAnalysisTask`
- `PKParameterTask`
- `PopulationTask`
- `FullPathDisplayResolver`
- `SensitivityAnalysisRunner`
- `SimulationResultsTask`
- `SimulationPersister`
- `SimulationRunner`
- `ConcurrentSimulationRunner`
- *`DimensionTask` (already cached but not used everywhere!)*
- *`ContainerTask` (already cached)*

The problem of the `toUnit()` function was that it was creating new dimension tasks instead of loading it from cache!


So I would propose the following:

- Function `.getNetTask()` remains as is and returns a new tasks (for compatibility reasons)
- Function `.getNetTaskFromChache()` returns a cached version. Cached versions should be stored in the `tasks-env` environment
- Remove functions `.getContainerTask()` and `.getDimensionTask()`. Instead, everywhere where they were used use `.getNetTaskFromChache()`. This way we will not have to create new functions every time we decide to cache a task.
- Make sure that the cached tasks are NEVER created with `.getNetTask()` but always with `.getNetTaskFromChache()`.
- I propose to also cache the tasks `DataRepositoryTask` and `DataImporterTask`.  For other tasks, its up to @msevestre to decide.