@Felixmil 2023-04-18
All @PavelBal suggestions were implemented in last commits.  This is ready for review @msevestre.